+++ src/inspect_ai/log/_recorders/buffer/types.py
@@ -62,7 +62,7 @@ class SampleBuffer(abc.ABC):
         ...
 
     @abc.abstractmethod
-    def get_sample_data(
+    async def get_sample_data(
         self,
         id: str | int,
         epoch: int,